fix: updated balance

+++ controllers/transactionController.js
@@ -18,7 +18,8 @@ export const getTransactions = async (req, res) => {
         'user category detail type amount date'
       )
         .populate('user', 'name')
-        .populate('category', 'name');
+        .populate('category', 'name')
+        .sort({ created_at: -1 });
       return res.status(200).json(transactions);
     }
     const transactions = await Transaction.find(
@@ -26,7 +27,8 @@ export const getTransactions = async (req, res) => {
       'user category detail type amount date'
     )
       .populate('user', 'name')
-      .populate('category', 'name');
+      .populate('category', 'name')
+      .sort({ created_at: -1 });
     res.status(200).json(transactions);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -75,6 +77,13 @@ export const createTransaction = async (req, res) => {
       amount,
       date,
     });
+    if (type === 'income') {
+      const balance = authUser[0].balance + amount;
+      await User.updateOne({ refresh_token: refreshToken }, { balance });
+    } else if (type === 'expense') {
+      const balance = authUser[0].balance - amount;
+      await User.updateOne({ refresh_token: refreshToken }, { balance });
+    }
     res.status(201).json({
       message: 'Transaction created successfully',
       data: createdTransaction,
@@ -92,11 +101,27 @@ export const updateTransaction = async (req, res) => {
     return res.status(404).json({ message: 'Transaction not found' });
   }
   try {
+    const refreshToken = req.cookies.refreshToken;
+    if (!refreshToken) return res.sendStatus(401);
+    const authUser = await User.find({ refresh_token: refreshToken });
+    if (!authUser[0]) return res.sendStatus(403);
     const updatedTransaction = await Transaction.findById(req.params.id);
     if (!updatedTransaction) {
       return res.status(404).json({ message: 'Transaction not found' });
     }
-    await Transaction.updateOne({ _id: req.params.id }, { $set: req.body });
+    const { type, amount } = updatedTransaction;
+    const updated = await Transaction.findByIdAndUpdate(
+      { _id: req.params.id },
+      { $set: req.body }
+    );
+    const updatedBalance = req.body.amount - amount;
+    if (type === 'income') {
+      const balance = authUser[0].balance + updatedBalance;
+      await User.updateOne({ refresh_token: refreshToken }, { balance });
+    } else if (type === 'expense') {
+      const balance = authUser[0].balance - updatedBalance;
+      await User.updateOne({ refresh_token: refreshToken }, { balance });
+    }
     res.status(200).json({ message: 'Transaction updated successfully' });
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -111,9 +136,21 @@ export const deleteTransaction = async (req, res) => {
     return res.status(404).json({ message: 'Transaction not found' });
   }
   try {
+    const refreshToken = req.cookies.refreshToken;
+    if (!refreshToken) return res.sendStatus(401);
+    const authUser = await User.find({ refresh_token: refreshToken });
+    if (!authUser[0]) return res.sendStatus(403);
     const transaction = await Transaction.findById(req.params.id);
     if (!transaction) {
       return res.status(404).json({ message: 'Transaction not found' });
+    }
+    const { type, amount } = transaction;
+    if (type === 'income') {
+      const balance = authUser[0].balance - amount;
+      await User.updateOne({ refresh_token: refreshToken }, { balance });
+    } else if (type === 'expense') {
+      const balance = authUser[0].balance + amount;
+      await User.updateOne({ refresh_token: refreshToken }, { balance });
     }
     await Transaction.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: 'Transaction deleted successfully' });
